Move retrieve_recovery_packets into handle_recovery
